# docs/api/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Whatboo API",
    "version": "2026.06.30",
    "description": "Contrato operativo para la API multi-tenant de Whatboo. Todas las rutas privadas usan el prefijo global /api/v1, JWT bearer y cabecera tenant salvo que se indique lo contrario."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [],
      "tenantSlug": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login de usuario tenant-aware",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesion iniciada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Solicitar reset de password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          }
        },
        "responses": {
          "202": { "description": "Solicitud aceptada" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Completar reset de password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Password actualizado" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/register-tenant": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Crear tenant mediante token de onboarding",
        "security": [{ "onboardingToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterTenantRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Tenant creado" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Listar staff del tenant",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "name": "role", "in": "query", "schema": { "$ref": "#/components/schemas/UserRole" } },
          { "name": "isActive", "in": "query", "schema": { "type": "boolean" } },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/TakeQuery" },
          { "$ref": "#/components/parameters/SkipQuery" }
        ],
        "responses": { "200": { "description": "Usuarios paginados" } }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Crear usuario staff",
        "parameters": [{ "$ref": "#/components/parameters/TenantSlugHeader" }],
        "responses": { "201": { "description": "Usuario creado" } }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Ver usuario staff",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": { "200": { "description": "Usuario" } }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Actualizar usuario staff",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": { "200": { "description": "Usuario actualizado" } }
      }
    },
    "/platform/admin/dashboard": {
      "get": {
        "tags": ["Platform Admin"],
        "summary": "Metricas globales SaaS",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dashboard global",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlatformDashboard" }
              }
            }
          }
        }
      }
    },
    "/platform/admin/restaurants": {
      "get": {
        "tags": ["Platform Admin"],
        "summary": "Listar restaurantes de plataforma",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Restaurantes paginados",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedPlatformRestaurants" }
              }
            }
          }
        }
      }
    },
    "/platform/admin/restaurants/{id}": {
      "get": {
        "tags": ["Platform Admin"],
        "summary": "Ver restaurante de plataforma",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Detalle de restaurante",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlatformRestaurantDetail" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Platform Admin"],
        "summary": "Actualizar restaurante de plataforma",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePlatformRestaurantRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restaurante actualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlatformRestaurantListItem" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "tenantSlug": { "type": "apiKey", "in": "header", "name": "X-Tenant-Slug" },
      "onboardingToken": { "type": "apiKey", "in": "header", "name": "X-Onboarding-Token" }
    },
    "parameters": {
      "TenantSlugHeader": {
        "name": "X-Tenant-Slug",
        "in": "header",
        "required": false,
        "schema": { "type": "string" }
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": { "type": "string" }
      },
      "TakeQuery": {
        "name": "take",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
      },
      "SkipQuery": {
        "name": "skip",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 0 }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Solicitud invalida",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "No autenticado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Operacion no permitida",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password", "restaurantSlug"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 1 },
          "restaurantSlug": { "type": "string", "minLength": 1 }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["accessToken"],
        "properties": {
          "accessToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["email", "restaurantSlug"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "restaurantSlug": { "type": "string" }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["token", "password"],
        "properties": {
          "token": { "type": "string" },
          "password": { "type": "string", "minLength": 12 }
        }
      },
      "RegisterTenantRequest": {
        "type": "object",
        "required": ["restaurantName", "slug", "ownerEmail", "ownerPassword"],
        "properties": {
          "restaurantName": { "type": "string" },
          "slug": { "type": "string" },
          "ownerEmail": { "type": "string", "format": "email" },
          "ownerPassword": { "type": "string", "minLength": 12 }
        }
      },
...
      "UserRole": {
        "type": "string",
        "enum": ["OWNER", "MANAGER", "STAFF", "ADMIN", "PLATFORM_ADMIN"]
      },
      "RestaurantStatus": {
        "type": "string",
        "enum": ["ACTIVE", "INACTIVE", "SUSPENDED"]
      },
      "BillingPlan": {
        "type": "string",
        "enum": ["FREE", "STARTER", "PRO", "ENTERPRISE"]
      },
      "BillingStatus": {
        "type": "string",
        "enum": ["TRIAL", "ACTIVE", "PAST_DUE", "SUSPENDED", "CANCELLED"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": ["string", "null"] },
          "lastName": { "type": ["string", "null"] },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "isActive": { "type": "boolean" }
        }
      },
      "PlatformDashboard": {
        "type": "object",
        "required": ["totals", "metricsByTenant"],
        "properties": {
          "totals": {
            "type": "object",
            "required": ["restaurants", "restaurantsActive", "restaurantsInactive", "restaurantsSuspended", "billingTrial", "billingActive", "billingPastDue", "billingSuspended", "users", "reservations"],
            "properties": {
              "restaurants": { "type": "integer" },
              "restaurantsActive": { "type": "integer" },
              "restaurantsInactive": { "type": "integer" },
              "restaurantsSuspended": { "type": "integer" },
              "billingTrial": { "type": "integer" },
              "billingActive": { "type": "integer" },
              "billingPastDue": { "type": "integer" },
              "billingSuspended": { "type": "integer" },
              "users": { "type": "integer" },
              "reservations": { "type": "integer" }
            }
          },
          "metricsByTenant": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PlatformTenantMetric" }
          }
        }
      },
      "PlatformTenantMetric": {
        "type": "object",
        "required": ["id", "slug", "name", "status", "users", "customers", "reservations", "tables"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "status": { "$ref": "#/components/schemas/RestaurantStatus" },
          "users": { "type": "integer" },
          "customers": { "type": "integer" },
          "reservations": { "type": "integer" },
          "tables": { "type": "integer" }
        }
      },
      "PlatformRestaurantListItem": {
        "type": "object",
        "required": ["id", "slug", "name", "timezone", "currency", "locale", "status", "billingPlan", "billingStatus", "createdAt", "updatedAt", "_count"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "legalName": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"], "format": "email" },
          "phone": { "type": ["string", "null"] },
          "city": { "type": ["string", "null"] },
          "country": { "type": ["string", "null"] },
          "timezone": { "type": "string" },
          "currency": { "type": "string" },
          "locale": { "type": "string" },
          "maxCapacity": { "type": ["integer", "null"] },
          "status": { "$ref": "#/components/schemas/RestaurantStatus" },
          "billingPlan": { "$ref": "#/components/schemas/BillingPlan" },
          "billingStatus": { "$ref": "#/components/schemas/BillingStatus" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "_count": {
            "type": "object",
            "properties": {
              "users": { "type": "integer" },
              "customers": { "type": "integer" },
              "reservations": { "type": "integer" },
              "tables": { "type": "integer" }
            }
          }
        }
      },
      "PlatformRestaurantDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/PlatformRestaurantListItem" },
          {
            "type": "object",
            "properties": {
              "taxId": { "type": ["string", "null"] },
              "website": { "type": ["string", "null"] },
              "addressLine1": { "type": ["string", "null"] },
              "allowWaitlist": { "type": "boolean" },
              "autoConfirm": { "type": "boolean" },
              "billingEmail": { "type": ["string", "null"], "format": "email" },
              "billingCustomerReference": { "type": ["string", "null"] },
              "trialEndsAt": { "type": ["string", "null"], "format": "date-time" },
              "currentPeriodEndsAt": { "type": ["string", "null"], "format": "date-time" },
              "defaultReservationDuration": { "type": "integer" },
              "slotIntervalMinutes": { "type": "integer" },
              "bufferTimeMinutes": { "type": "integer" },
              "whatsappAccounts": { "type": "array", "items": { "type": "object" } },
              "openingHours": { "type": "array", "items": { "type": "object" } },
              "capacityRules": { "type": "array", "items": { "type": "object" } }
            }
          }
        ]
      },
      "PaginatedPlatformRestaurants": {
        "type": "object",
        "required": ["data", "total", "take", "skip"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/PlatformRestaurantListItem" } },
          "total": { "type": "integer" },
          "take": { "type": "integer" },
          "skip": { "type": "integer" }
        }
      },
      "UpdatePlatformRestaurantRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "legalName": { "type": ["string", "null"] },
          "taxId": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"], "format": "email" },
          "phone": { "type": ["string", "null"] },
          "website": { "type": ["string", "null"] },
          "addressLine1": { "type": ["string", "null"] },
          "city": { "type": ["string", "null"] },
          "country": { "type": ["string", "null"] },
          "timezone": { "type": "string" },
          "currency": { "type": "string" },
          "locale": { "type": "string" },
          "maxCapacity": { "type": ["integer", "null"] },
          "allowWaitlist": { "type": "boolean" },
          "autoConfirm": { "type": "boolean" },
          "status": { "$ref": "#/components/schemas/RestaurantStatus" },
          "billingPlan": { "$ref": "#/components/schemas/BillingPlan" },
          "billingStatus": { "$ref": "#/components/schemas/BillingStatus" },
          "billingEmail": { "type": ["string", "null"], "format": "email" },
          "billingCustomerReference": { "type": ["string", "null"] },
          "trialEndsAt": { "type": ["string", "null"], "format": "date-time" },
          "currentPeriodEndsAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "error": { "type": "string" }
        }
      }
    }
  }
}